agent_test/requirements.txt
# MCP服务依赖
langchain-openai
langchain-community
langchain-mcp-adapters
mcp
python-dotenv

# 智能体依赖
langchain-openai
langchain-community
python-dotenv
langchain